Make sure import script is consistent across OS revs
Previous versions of Perl would not try to use UTF-8 on the scripts, but
the newer versions do. Disable this to make sure the output is
repeatable across Perl revisions.

Also the output format of "file" had changed across revisions of the OS,
so make sure we specify a flag that works to print out the
machine-readable MIME encoding.

Change-Id: I4213f0e57a16c030383be0a5f5ceb6b1e6e976e5